refactor: 내비게이션 시스템 분리 및 Index 페이지 정리
