+++ src/components/organism/NavigationSystem.jsx
@@ -1,0 +1,27 @@
+// components/organisms/NavigationSystem.jsx
+import React, { useState } from 'react';
+import NavBar from './NavBar';
+import HamburgerMenuButton from '../molecules/HamburgerMenuButton';
+import BackgroundOverlay from '../molecules/BackgroundOverlay';
+
+function NavigationSystem() {
+  const [isNavOpen, setIsNavOpen] = useState(false);
+
+  const toggleNavBar = () => {
+    setIsNavOpen((prev) => !prev);
+  };
+
+  const closeNavBar = () => {
+    setIsNavOpen(false);
+  };
+
+  return (
+    <>
+      <HamburgerMenuButton isVisible={!isNavOpen} onClick={toggleNavBar} />
+      <NavBar isOpen={isNavOpen} closeNavBar={closeNavBar} />
+      <BackgroundOverlay isOpen={isNavOpen} onClick={closeNavBar} />
+    </>
+  );
+}
+
+export default NavigationSystem;
