+++ src/pages/index.jsx
@@ -1,7 +1,6 @@
-import React, { useState } from 'react';
-import NavBar from "../components/NavBar";
+import React from 'react';
+import NavigationSystem from '../components/organism/NavigationSystem';
 import styled, { createGlobalStyle } from 'styled-components';
-import hamburgerIcon from "../assets/hamburger-menu.png"; // 햄버거 메뉴 아이콘 import
 
 const GlobalStyle = createGlobalStyle`
   * {
@@ -17,38 +16,6 @@ const GlobalStyle = createGlobalStyle`
   }
 `;
 
-const HamburgerButton = styled.button`
-  position: fixed;
-  top: 20px;
-  left: 20px;
-  width: 30px;
-  height: 30px;
-  background: transparent;
-  border: none;
-  cursor: pointer;
-  z-index: 1101;
-
-  opacity: ${({ isVisible }) => (isVisible ? 1 : 0)};
-  pointer-events: ${({ isVisible }) => (isVisible ? "auto" : "none")};
-  transition: opacity 0.4s ease; /* 페이드 효과 */
-
-  img {
-    width: 19px;
-    height: 18px;
-  }
-`;
-
-const Overlay = styled.div`
-  display: ${({ isOpen }) => (isOpen ? 'block' : 'none')};
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100vh;
-  background-color: rgba(0, 0, 0, 0.6);
-  z-index: 1050;
-`;
-
 const Container = styled.div`
   background-color: #000000;
   width: 100%;
@@ -60,33 +27,11 @@ const Container = styled.div`
 `;
 
 function Index() {
-  const [isNavOpen, setIsNavOpen] = useState(false);
-
-  const toggleNavBar = () => {
-    setIsNavOpen((prev) => !prev);
-  };
-
-  const closeNavBar = () => {
-    setIsNavOpen(false);
-  };
-
   return (
     <>
       <GlobalStyle />
       <Container>
-        {/* 조건부 렌더링 없애고 항상 버튼을 DOM에 둠 */}
-        <HamburgerButton
-          isVisible={!isNavOpen}
-          onClick={toggleNavBar}
-          aria-label="메뉴 열기"
-        >
-          <img src={hamburgerIcon} alt="햄버거 메뉴 아이콘" />
-        </HamburgerButton>
-        
-        {/* NavBar */}
-        <NavBar isOpen={isNavOpen} closeNavBar={closeNavBar} />
-        
-        <Overlay isOpen={isNavOpen} />
+        <NavigationSystem />
       </Container>
     </>
   );
